backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.23
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pydantic==2.5.2
-e ../../SNEWS_DB 